+++ src/Tests/Dfe.Complete.Application.Tests/QueryHandlers/Project/GetProjectHistoryByProjectIdQueryHandlerTests.cs
@@ -3,13 +3,12 @@ using DfE.CoreLibs.Testing.AutoFixture.Attributes;
 using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Application.Projects.Queries.GetProject;
 using Dfe.Complete.Domain.Entities;
-using Dfe.Complete.Domain.Interfaces.Repositories;
 using Dfe.Complete.Domain.ValueObjects;
 using NSubstitute;
 using AutoMapper;
+using Dfe.Complete.Application.Projects.Interfaces;
 using DfE.CoreLibs.Testing.AutoFixture.Customizations;
-using MockQueryable.NSubstitute;
-using NSubstitute.ExceptionExtensions;
+using MockQueryable;
 
 namespace Dfe.Complete.Application.Tests.QueryHandlers.Project
 {
@@ -18,26 +17,29 @@ namespace Dfe.Complete.Application.Tests.QueryHandlers.Project
         [Theory]
         [CustomAutoData(typeof(DateOnlyCustomization), typeof(OmitCircularReferenceCustomization))]
         public async Task Handle_ShouldReturnSuccess_WhenProjectExists(
-            [Frozen] ICompleteRepository<Domain.Entities.Project> mockProjectRepository,
+            [Frozen] IProjectReadRepository mockProjectRepository,
             [Frozen] IMapper mockMapper,
             GetProjectHistoryByProjectIdHandler handler,
-            Domain.Entities.Project project,
-            ProjectDto mappedProject
-        )
+            ProjectDto mappedProject)
         {
             // Arrange
             var validGuid = Guid.NewGuid();
             var query = new GetProjectHistoryByProjectIdQuery(validGuid.ToString());
 
-            project.Id = new ProjectId(validGuid);
-            project.Notes = new List<Note>
+            var projects = new List<Domain.Entities.Project>
             {
-                new Note { NotableType = "SignificantDateHistoryReason" },
-                new Note { NotableType = "OtherType" }
-            };
+                new ()
+                {
+                    Id = new ProjectId(validGuid),
+                    Notes = new List<Note>
+                    {
+                        new Note { NotableType = "SignificantDateHistoryReason" }
+                    },
+                }
+            }.AsQueryable().BuildMock();
 
-            var mockDbSet = new List<Domain.Entities.Project> { project }.AsQueryable().BuildMockDbSet();
-            mockProjectRepository.Query().Returns(mockDbSet);
+            mockProjectRepository.Projects.Returns(projects);
+
             mockMapper.Map<ProjectDto>(Arg.Any<Domain.Entities.Project>()).Returns(mappedProject);
 
             // Act
@@ -48,57 +50,39 @@ namespace Dfe.Complete.Application.Tests.QueryHandlers.Project
             Assert.NotNull(result.Value);
             Assert.Equal(mappedProject, result.Value);
         }
-
-
-        [Theory]
-        [CustomAutoData(typeof(DateOnlyCustomization), typeof(OmitCircularReferenceCustomization))]
-        public async Task Handle_ShouldReturnFailure_WhenExceptionThrown(
-            [Frozen] ICompleteRepository<Domain.Entities.Project> mockProjectRepository,
-            GetProjectHistoryByProjectIdHandler handler
-        )
-        {
-            // Arrange
-            var query = new GetProjectHistoryByProjectIdQuery(Guid.NewGuid().ToString());
-            mockProjectRepository.Query().Throws(new Exception("DB Error"));
-
-            // Act
-            var result = await handler.Handle(query, default);
-
-            // Assert
-            Assert.False(result.IsSuccess);
-            Assert.Contains("DB Error", result.Error);
-        }
-        
         
         [Theory]
         [CustomAutoData(typeof(DateOnlyCustomization), typeof(OmitCircularReferenceCustomization))]
         public async Task Handle_ShouldReturnFailure_WhenSignificantDateHasNoUser(
-            [Frozen] ICompleteRepository<Domain.Entities.Project> mockProjectRepository,
-            GetProjectHistoryByProjectIdHandler handler,
-            Domain.Entities.Project project
+            [Frozen] IProjectReadRepository mockProjectRepository,
+            GetProjectHistoryByProjectIdHandler handler
         )
         {
             // Arrange
             var validGuid = Guid.NewGuid();
             var query = new GetProjectHistoryByProjectIdQuery(validGuid.ToString());
-
-            project.Id = new ProjectId(validGuid);
-            project.Notes = new List<Note>
+            
+            var projects = new List<Domain.Entities.Project>
             {
-                new Note { NotableType = "SignificantDateHistoryReason" }
-            };
-
-            project.SignificantDateHistories = new List<SignificantDateHistory>
-            {
-                new SignificantDateHistory
+                new ()
                 {
-                    User = null,
-                    Reason = new SignificantDateHistoryReason()
+                    Id = new ProjectId(validGuid),
+                    Notes = new List<Note>
+                    {
+                        new Note { NotableType = "SignificantDateHistoryReason" }
+                    },
+                    SignificantDateHistories = new List<SignificantDateHistory>
+                    {
+                        new()
+                        {
+                            User = null,
+                            Reason = new SignificantDateHistoryReason()
+                        }
+                    }
                 }
-            };
+            }.AsQueryable().BuildMock();
 
-            var mockDbSet = new List<Domain.Entities.Project> { project }.AsQueryable().BuildMockDbSet();
-            mockProjectRepository.Query().Returns(mockDbSet);
+            mockProjectRepository.Projects.Returns(projects);
 
             // Act
             var result = await handler.Handle(query, default);